% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{cola_year}
\alias{cola_year}
\title{The year of the benefit increase in effect}
\usage{
cola_year(year, month)
}
\arguments{
\item{year, month}{A benefit month.}
}
\value{
The year of the latest benefit increase in effect.
}
\description{
Increases take effect in December from 1983, and in June 1974-1982.
}
\examples{
cola_year(2026, c(11, 12))
}
