% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claiming.R
\name{eligibility_year}
\alias{eligibility_year}
\title{Eligibility year}
\usage{
eligibility_year(
  birth_year,
  birth_month,
  birth_day = 15,
  onset_year = NULL,
  death_year = NULL
)
}
\arguments{
\item{birth_year, birth_month, birth_day}{Date of birth.}

\item{onset_year, death_year}{Optional year of disability onset or death.}
}
\value{
Eligibility years.
}
\description{
The year of attaining 62, or of disability onset or death if earlier.
}
\examples{
eligibility_year(1960, 6)
}
