% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{family_max}
\alias{family_max}
\title{Maximum family benefit}
\usage{
family_max(pia_elig, elig_year, policy = current_law())
}
\arguments{
\item{pia_elig}{PIA at eligibility.}

\item{elig_year}{Eligibility year.}

\item{policy}{A \code{\link[=current_law]{current_law()}} policy.}
}
\value{
Family maximums.
}
\description{
The family maximum at eligibility, before COLAs, for retirement and
survivor cases. Disability uses \code{\link[=di_family_max]{di_family_max()}}.
}
