% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claiming.R
\name{monthly_benefit}
\alias{monthly_benefit}
\title{Monthly benefit payable}
\usage{
monthly_benefit(pia, factor, benefit_year, benefit_month = 12)
}
\arguments{
\item{pia}{PIA in the benefit month.}

\item{factor}{Reduction or credit factor.}

\item{benefit_year, benefit_month}{The benefit month.}
}
\value{
Dollars.
}
\description{
factor x PIA, rounded down to the dime, then paid in whole dollars.
}
\examples{
monthly_benefit(1234.5, 0.7, 2030, 6)
}
