% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claiming.R
\name{normal_retirement_age}
\alias{normal_retirement_age}
\title{Normal retirement age}
\usage{
normal_retirement_age(
  birth_year,
  birth_month,
  birth_day = 15,
  policy = current_law()
)
}
\arguments{
\item{birth_year}{Date of birth.}

\item{birth_month}{Date of birth.}

\item{birth_day}{Date of birth.}

\item{policy}{A \code{\link[=current_law]{current_law()}} policy.}
}
\value{
The normal (full) retirement age in months.
}
\description{
Normal retirement age
}
\examples{
normal_retirement_age(1960, 6) / 12
}
